+++ app/src/main/java/ru/practicum/android/diploma/di/ViewModelModule.kt
@@ -2,8 +2,8 @@ package ru.practicum.android.diploma.di
 
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
-import ru.practicum.android.diploma.feature.vacancy.presentation.VacancyViewModel
 import ru.practicum.android.diploma.feature.favorite.presentation.FavoriteFragmentViewModel
+import ru.practicum.android.diploma.feature.vacancy.presentation.VacancyViewModel
 
 val viewModelModule = module {
