+++ app/src/main/java/ru/practicum/android/diploma/feature/vacancy/presentation/VacancyState.kt
@@ -1,13 +1,9 @@
 package ru.practicum.android.diploma.feature.vacancy.presentation
 
-import androidx.annotation.StringRes
 import ru.practicum.android.diploma.feature.vacancy.domain.model.VacancyDetail
 
 sealed class VacancyState {
     object Loading : VacancyState()
     data class Content(val content: VacancyDetail) : VacancyState()
     data class Error(val errorMessage: String) : VacancyState()
-
-    data class Empty(@StringRes val message: Int) : VacancyState()
-
 }
